requirements.txt
streamlit
numpy
pandas
scipy
dataclasses
PyYAML
plotly
requests
json
re